+++ tests/unit/data/image/test_image_properties.py
