+++ tests/unit/data/image/test_intensity_image.py
