+++ tests/unit/data/image/test_segmentation_image.py
