+++ tests/unit/fileio/crawling/test_subject_file_crawler.py
