+++ tests/unit/fileio/series_info/test_dicom_series_image_info.py
