+++ tests/unit/fileio/series_info/test_dicom_series_registration_info.py
