+++ tests/unit/fileio/series_info/test_intensity_file_series_info.py
