+++ tests/unit/fileio/series_info/test_segmentation_file_series_info.py
