+++ tests/unit/fileio/series_info/test_series_info.py
